Extra Task 2 - Search

+++ backend-assignment/src/main/java/com/example/backend_assignment/repository/MetarRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.backend_assignment.model.Metar;
 
-public interface MetarRepository extends JpaRepository<Metar, Integer> {
+public interface MetarRepository extends JpaRepository<Metar, Long> {
 	Optional<Metar> findBySubscriptionId(Long id);
 }
